Clean up LIF_net stuff
