+++ examples/bioRxiv_2022.02.28.482256/Fit_LIF_net.py
@@ -1,77 +1,47 @@
 #!/usr/bin/env python
 # coding: utf-8
-
-# In[ ]:
-
-
-from pymoo.factory import get_visualization
 from pymoo.factory import get_decision_making, get_reference_directions
-from pymoo.visualization.scatter import Scatter
 from pymoo.model.callback import Callback
 from pymoo.optimize import minimize
 from pymoo.algorithms.nsga2 import NSGA2
 from pymoo.model.problem import Problem
-from nest import raster_plot
 import nest
-import pandas as pd
 import hashlib
 import json
-from copy import deepcopy
-import scipy.stats as st
-# import example_network_parameters as params
 from example_network_parameters import (networkParameters, population_names,
                                         population_sizes)
+import example_network_methods as methods
 import scipy.signal as ss
 import h5py
-from matplotlib.gridspec import GridSpec
 import matplotlib.pyplot as plt
 import numpy as np
 from parameters import ParameterSpace, ParameterSet
 import os
-from time import time
-# get_ipython().run_line_magic('matplotlib', 'inline')
+import LIF_net
 
 
-# In[ ]:
-
-
-# In[ ]:
-
-
+# print only NEST errors
 nest.set_verbosity('M_ERROR')
 
 
-# In[ ]:
-
-
+# Load parameter spaces
 PS0 = ParameterSpace('PS0.txt')
 PS1 = ParameterSpace('PS1.txt')
-# PS2 = ParameterSpace('PS2.txt')
 
-
-# In[ ]:
-
-
+# simulation parameters
 TRANSIENT = 2000
 dt = networkParameters['dt']
 tstop = networkParameters['tstop']
-tau = 100  # time lag relative to spike for kernel predictions
 
 
-# In[ ]:
-
-
-# plt.mlab.psd/csd settings
+# scipy.signal.psd settings for power spectra
 Fs = 1000 / dt
-NFFT = 1024 * 2
-noverlap = 768 * 2
+NFFT = 2048
+noverlap = 1536
 detrend = 'constant'
+cutoff = 200
 
-
-# In[ ]:
-
-
-# figure out which real LFP to compare with
+# figure out which ground truth data to compare with
 for pset in PS1.iter_inner():
     weight_EE = pset['weight_EE']
     weight_IE = pset['weight_IE']
@@ -92,43 +62,6 @@ for pset in PS1.iter_inner():
 print(f'comparing with ground truth dataset: {OUTPUTPATH_REAL}')
 
 
-# In[ ]:
-
-
-def get_spike_rate(times):
-    bins = (np.arange(TRANSIENT / dt, tstop / dt + 1)
-            * dt - dt / 2)
-    hist, _ = np.histogram(times, bins=bins)
-    return bins, hist.astype(float) / dt
-
-
-# In[ ]:
-
-
-def get_mean_spike_rate(times):
-    times = times[times >= TRANSIENT]
-    return times.size / (tstop - TRANSIENT) * 1000
-
-
-# In[ ]:
-
-
-def get_psd(nu, cutoff=200):
-    freqs, Pxx = ss.welch(nu, fs=Fs, nperseg=NFFT,
-                          noverlap=noverlap, detrend=detrend)
-    return freqs[freqs <= cutoff], Pxx[freqs <= cutoff]
-
-
-# In[ ]:
-
-
-def zscore(x):
-    return (x - x.mean()) / x.std()
-
-
-# In[ ]:
-
-
 # mean firing rates, rates, rate spectras of "real" network populations
 mean_nu_X = dict()
 nu_X = dict()
@@ -136,242 +69,13 @@ psd_X = dict()
 with h5py.File(os.path.join(OUTPUTPATH_REAL, 'spikes.h5'), 'r') as f:
     for i, X in enumerate(population_names):
         times = np.concatenate(f[X]['times'])
-        mean_nu_X[X] = get_mean_spike_rate(times) / population_sizes[i]
-        bins, nu_X[X] = get_spike_rate(times)
-        freqs, psd_X[X] = get_psd(nu_X[X])
-
-mean_nu_X
-
-
-# In[ ]:
-
-
-# plot firing rate time series of "real" network
-inds = (bins[:-1] >= TRANSIENT) & (bins[:-1] <= TRANSIENT + 100)
-fig, axes = plt.subplots(2, 1, sharex=True, sharey=True, figsize=(16, 9))
-with h5py.File(os.path.join(OUTPUTPATH_REAL, 'spikes.h5'), 'r') as f:
-    for i, (X, N_X) in enumerate(zip(population_names, population_sizes)):
-        axes[i].step(bins[:-1][inds], nu_X[X][inds])
-        axes[i].set_title(r'$\nu_{%s}(t)$' % X)
-        axes[i].set_ylabel(r'$\nu_X$ (# spikes/s)')
-axes[1].set_xlabel('$t$ (ms)')
-
-
-# In[ ]:
-
-
-# firing rate spectra
-fig, ax = plt.subplots(1, 1, figsize=(8, 8))
-for X in population_names:
-    ax.plot(freqs, psd_X[X], label=X)
-ax.set_xlabel('$f$ (Hz)', labelpad=0)
-ax.set_ylabel(r'PSD$_\nu$ (s$^{-2}$/Hz)')
-ax.legend()
-
-
-# In[ ]:
-
-
-# create NEST LIF network
-class Network(object):
-    """Class implementing a LIF network"""
-
-    def __init__(self,
-                 X=['E', 'I'],
-                 N_X=[8192, 1024],
-                 C_m_X=[269., 100.],
-                 tau_m_X=[10., 10.],
-                 E_L_X=[-65., -65.],
-                 C_YX=[[0.5, 0.5], [0.5, 0.5]],
-                 J_YX=[[1.5, 1.5], [-20., -15.]],
-                 delay_YX=[[3., 3.], [3.0, 3.0]],
-                 tau_syn_YX=[[0.5, 0.5], [0.5, 0.5]],
-                 n_ext=[450, 160],
-                 nu_ext=40.,
-                 J_ext=27.6,
-                 model='iaf_psc_exp',
-                 dt=2**-4,
-                 verbose=False,
-                 **kwargs
-                 ):
-        self.X = X
-        self.N_X = N_X
-        self.C_m_X = C_m_X
-        self.tau_m_X = tau_m_X
-        self.E_L_X = E_L_X
-        self.C_YX = C_YX
-        self.J_YX = J_YX
-        self.delay_YX = delay_YX
-        self.tau_syn_YX = tau_syn_YX
-        self.n_ext = n_ext
-        self.nu_ext = nu_ext
-        self.J_ext = J_ext
-        self.model = model
-        self.dt = dt
-        self.verbose = verbose
-
-        self._create()
-        self._connect()
-
-    def _create(self):
-        """Create network nodes and connections"""
-
-        nest.ResetKernel()
-        nest.SetKernelStatus(
-            dict(
-                local_num_threads=32,
-                resolution=self.dt,
-                tics_per_ms=1000 / dt))
-
-        if self.verbose:
-            print('creating...')
-        # neurons
-        self.neurons = {}
-        for (
-                X,
-                N,
-                C_m,
-                tau_m,
-                E_L,
-                (tau_syn_ex,
-                 tau_syn_in)) in zip(
-                self.X,
-                self.N_X,
-                self.C_m_X,
-                self.tau_m_X,
-                self.E_L_X,
-                self.tau_syn_YX):
-            params = dict(
-                C_m=C_m,
-                tau_m=tau_m,
-                E_L=E_L,
-                V_reset=E_L,
-                tau_syn_ex=tau_syn_ex,
-                tau_syn_in=tau_syn_in
-            )
-            self.neurons[X] = nest.Create(self.model, N, params)
-
-        # poisson generators
-        self.poisson = {}
-        for X, n_ext in zip(self.X, self.n_ext):
-            self.poisson[X] = nest.Create(
-                'poisson_generator', 1, dict(
-                    rate=self.nu_ext * n_ext))
-
-        # spike recorders
-        self.spike_recorders = {}
-        for X in self.X:
-            self.spike_recorders[X] = nest.Create('spike_recorder', 1)
-
-    def _connect(self):
-        if self.verbose:
-            print('connecting...')
-        for i, X in enumerate(self.X):
-            # recurrent connections
-            for j, Y in enumerate(self.X):
-                if self.verbose:
-                    print(f'connecting {X} to {Y}...')
-                conn_spec = dict(
-                    rule='pairwise_bernoulli',
-                    p=self.C_YX[i][j],
-                )
-                syn_spec = dict(
-                    synapse_model='static_synapse',
-                    weight=nest.math.redraw(
-                        nest.random.normal(
-                            mean=self.J_YX[i][j],
-                            std=abs(self.J_YX[i][j]) * 0.1,
-                        ),
-                        min=0. if self.J_YX[i][j] >= 0 else np.NINF,
-                        max=np.Inf if self.J_YX[i][j] >= 0 else 0.,
-                    ),
-
-                    delay=nest.math.redraw(
-                        nest.random.normal(
-                            mean=self.delay_YX[i][j],
-                            std=self.delay_YX[i][j] * 0.5,
-                        ),
-                        min=0.3,
-                        max=np.Inf,
-                    )
-                )
-
-                nest.Connect(
-                    self.neurons[X],
-                    self.neurons[Y],
-                    conn_spec,
-                    syn_spec)
-
-            # poisson generators
-            if self.verbose:
-                print(f'connecting poisson_generator[{X}] to {X}...')
-            nest.Connect(
-                self.poisson[X],
-                self.neurons[X],
-                'all_to_all',
-                dict(
-                    weight=self.J_ext))
-
-            # recorders
-            if self.verbose:
-                print(f'connecting spike_recorder[{X}] to {X}...')
-            nest.Connect(self.neurons[X], self.spike_recorders[X])
-
-    def simulate(self, tstop=6000):
-        """Instantiate and run simulation"""
-        if self.verbose:
-            print('simulating...')
-        nest.Simulate(tstop)
-        if self.verbose:
-            print('done!')
-
-
-# In[ ]:
-
-
-'''
-# test Network class
-tic = time()
-net = Network()
-net.simulate(tstop=tstop + dt)
-toc = time()
-print('simulation time:', toc-tic)
-'''
-
-'''
-# In[ ]:
-
-
-# mean firing rates of "real" network populations
-lif_mean_nu_X = dict()  # mean spike rates
-lif_nu_X = dict()  # binned firing rate
-lif_psd_X = dict()
-for i, X in enumerate(population_names):
-    times = nest.GetStatus(net.spike_recorders[X])[0]['events']['times']
-    times = times[times >= TRANSIENT]
-
-    lif_mean_nu_X[X] = get_mean_spike_rate(times)
-    _, lif_nu_X[X] = get_spike_rate(times)
-    _, lif_psd_X[X] = get_psd(lif_nu_X[X])
-
-lif_mean_nu_X
-
-
-# In[ ]:
-
-
-# power spectra
-fig, ax = plt.subplots(1, 1, figsize=(8, 8))
-for X in population_names:
-    ax.plot(freqs, psd_X[X], label=X)
-    ax.plot(freqs, lif_psd_X[X], label=X)
-ax.set_xlabel('$f$ (Hz)', labelpad=0)
-ax.set_ylabel(r'PSD$_\nu$ (s$^{-2}$/Hz)')
-ax.legend()
-
-
-# In[ ]:
-'''
+        mean_nu_X[X] = LIF_net.get_mean_spike_rate(
+            times, TRANSIENT=TRANSIENT, tstop=tstop) / population_sizes[i]
+        bins, nu_X[X] = LIF_net.get_spike_rate(
+            times, TRANSIENT=TRANSIENT, tstop=tstop, dt=dt)
+        freqs, psd_X[X] = LIF_net.get_psd(
+            nu_X[X], Fs=Fs, NFFT=NFFT, noverlap=noverlap, detrend=detrend,
+            cutoff=cutoff)
 
 
 def get_lif_data(x=[1.5, 1.5, -20., -15., 27.6, 3.0, 3.0, 3.0, 3.0, 269, 100]):
@@ -391,8 +95,9 @@ def get_lif_data(x=[1.5, 1.5, -20., -15., 27.6, 3.0, 3.0, 3.0, 3.0, 269, 100]):
         J_ext=x[4],
         model='iaf_psc_exp',
         dt=2**-4,
+        local_num_threads=32,
     )
-    net = Network(**params)
+    net = LIF_net.Network(**params)
     net.simulate(tstop=tstop + dt)
 
     psd_X = []
@@ -402,36 +107,30 @@ def get_lif_data(x=[1.5, 1.5, -20., -15., 27.6, 3.0, 3.0, 3.0, 3.0, 269, 100]):
         times = nest.GetStatus(net.spike_recorders[X])[0]['events']['times']
         times = times[times >= TRANSIENT]
 
-        mean_nu_X += [get_mean_spike_rate(times) / population_sizes[i]]
-        _, lif_nu_X = get_spike_rate(times)
+        mean_nu_X += [LIF_net.get_mean_spike_rate(times,
+                                          TRANSIENT=TRANSIENT,
+                                          tstop=tstop) / population_sizes[i]]
+        _, lif_nu_X = LIF_net.get_spike_rate(
+            times, TRANSIENT=TRANSIENT, tstop=tstop, dt=dt)
         nu_X += [lif_nu_X]
-        _, lif_psd_X = get_psd(lif_nu_X)
+        _, lif_psd_X = LIF_net.get_psd(
+            lif_nu_X, Fs=Fs, NFFT=NFFT, noverlap=noverlap, detrend=detrend,
+            cutoff=cutoff)
         psd_X += [lif_psd_X]
 
     return mean_nu_X, nu_X, psd_X
-
-
-'''
-# In[ ]:
-
-
-lif_mean_nu_X, lif_nu_X, lif_psd_X = get_lif_data(
-    x=[1.5, 1.5, -20., -15., 27.6, 3.0, 3.0, 3.0, 3.0])
-
-
-# In[ ]:
-'''
 
 
 class LifNet(Problem):
     """inherited pymoo Problem class required for minimizer"""
 
     def __init__(self):
-        super().__init__(n_var=11,
-                         n_obj=4,
-                         n_constr=0,
-                         xl=[1.1, 1.5, -25., -14., 28, 1.0, 1.0, 1.0, 1.0, 270, 100],
-                         xu=[1.8, 2.1, -18., -8., 32, 4.0, 4.0, 4.0, 4.0, 310, 120])
+        super().__init__(
+            n_var=11,
+            n_obj=4,
+            n_constr=0,
+            xl=[1.1, 1.5, -25., -14., 28, 1.0, 1.0, 1.0, 1.0, 270, 100],
+            xu=[1.8, 2.1, -18., -8., 32, 4.0, 4.0, 4.0, 4.0, 310, 120])
 
     def _evaluate(self, x, out, *args, **kwargs):
         f0 = []
@@ -452,9 +151,6 @@ class LifNet(Problem):
         out['F'] = np.column_stack([f0, f1, f2, f3])
 
 
-# In[ ]:
-
-
 class MyCallback(Callback):
 
     def __init__(self) -> None:
@@ -465,9 +161,6 @@ class MyCallback(Callback):
     def notify(self, algorithm):
         self.data["F"].append(algorithm.pop.get("F"))
         self.data["X"].append(algorithm.pop.get("X"))
-
-
-# In[ ]:
 
 
 # minimize using NSGA-II multiobjective evolutionary algorithm
@@ -482,9 +175,6 @@ res = minimize(problem,
                verbose=True)
 
 
-# In[ ]:
-
-
 # Find a suitable parameter combination performing well with all features, from
 # https://pymoo.org/decision_making/index.html#Pseudo-Weights
 ref_dirs = get_reference_directions("das-dennis", 4, n_partitions=12)
@@ -493,31 +183,6 @@ F = res.F
 weights = np.array([0.25, 0.25, 0.25, 0.25])
 a, pseudo_weights = get_decision_making(
     "pseudo-weights", weights).do(res.F, return_pseudo_weights=True)
-'''
-plot = get_visualization("petal", bounds=(0, 0.5), reverse=True)
-plot.add(res.F[[a]])
-plot.show()
-'''
-
-# In[ ]:
-
-'''
-res.F[[a]], res.X[[a]][0]
-'''
-
-# In[ ]:
-
-'''
-# plot location in feature space
-plot = Scatter(figsize=(12, 12))
-for F in res.algorithm.callback.data["F"]:
-    plot.add(F, color="gray")
-plot.add(res.F, color="black")
-plot.add(res.F[[a]], color="red")
-plot.show()
-'''
-
-# In[ ]:
 
 
 # Store data to HDF5
@@ -528,21 +193,8 @@ with h5py.File('Fit_LIF_net.h5', 'w') as f:
     f["X_opt"] = res.X[[a]][0]
 
 
-# In[ ]:
-
-
-# pymoo messes up MPL rcParams
-# get_ipython().run_line_magic('matplotlib', 'inline')
-
-
-# In[ ]:
-
-
 # simulate network using updated params
 lif_mean_nu_X, lif_nu_X, lif_psd_X = get_lif_data(x=res.X[[a]][0])
-
-
-# In[ ]:
 
 
 # compare power spectra with updated params
@@ -553,10 +205,8 @@ for i, X in enumerate(population_names):
 ax.set_xlabel('$f$ (Hz)', labelpad=0)
 ax.set_ylabel(r'PSD$_\nu$ (s$^{-2}$/Hz)')
 ax.legend()
-fig.savefig('Fit_LIF_net_PSD.pdf', bbox_inches='tight')
-
-
-# In[ ]:
+fig.savefig(os.path.join('figures', 'Fit_LIF_net_PSD.pdf'),
+            bbox_inches='tight')
 
 
 # compare firing rate auto- and cross-correlation functions:
@@ -566,10 +216,10 @@ lag_inds = (lag >= -maxlag) & (lag <= maxlag)
 fig, axes = plt.subplots(2, 2, figsize=(16, 8), sharex=True, sharey=True)
 for i, X in enumerate(population_names):
     for j, Y in enumerate(population_names):
-        xcorr_mc = np.correlate(zscore(nu_X[X]),
-                                zscore(nu_X[Y]), 'same') / zscore(nu_X[X]).size
-        xcorr_lif = np.correlate(zscore(lif_nu_X[i]), zscore(
-            lif_nu_X[j]), 'same') / zscore(lif_nu_X[i]).size
+        xcorr_mc = np.correlate(methods.zscore(nu_X[X]), methods.zscore(
+            nu_X[Y]), 'same') / methods.zscore(nu_X[X]).size
+        xcorr_lif = np.correlate(methods.zscore(lif_nu_X[i]), methods.zscore(
+            lif_nu_X[j]), 'same') / methods.zscore(lif_nu_X[i]).size
 
         axes[i, j].plot(lag[lag_inds], xcorr_mc[lag_inds], label='mc')
         axes[i, j].plot(lag[lag_inds], xcorr_lif[lag_inds], label='lif')
@@ -584,7 +234,5 @@ for i, X in enumerate(population_names):
         axes[i, j].axis(axes[i, j].axis('tight'))
         axes[i, j].set_ylim(ymax=0.25)
         axes[i, j].legend()
-fig.savefig('Fit_LIF_net_CXY.pdf', bbox_inches='tight')
-
-
-# In[ ]:
+fig.savefig(os.path.join('figures', 'Fit_LIF_net_CXY.pdf'),
+            bbox_inches='tight')
